Add a note that this only works in Online Mode
Our Escalation Engineers asked me to submit this change because it was causing confusion for customers.  This is the note I am suggesting:

Note: This article applies only when using Outlook in online mode. Search folders do not sync, therefore search folders created in online mode will not appear in cached mode.