Update how-to-authenticate-an-imap-pop-smtp-application-by-using-oauth.md
Configurations described in [Configure a client application to access a web API](https://docs.microsoft.com/en-us/azure/active-directory/develop/quickstart-configure-app-access-web-apis) is not necessary. The applications might not be web apps, and basic configuration is already described in [Register an application with the Microsoft identity platform](https://docs.microsoft.com/en-us/azure/active-directory/develop/quickstart-register-app). Not only that, we cannot use permissions of Microsoft Graph. As described in "Get an access token" section, we need to acquire access tokens of Outlook. And we don't need to add permissions in advance, because we can acquire access tokens dynamically using MSAL or Microsoft Identity Platform v2 endpoint.