Fix namespace of Restriction (types -> messages)
It appears that Restriction is in the messages and not the types
namespace. While a Restriction element in the types namespace may be
included as a child of e.g. a FindFolder element, it is silently ignored
without raising an error. A Restriction element in the messages
namespace, on the other hand, is honored.