Update how-to-authenticate-an-imap-pop-smtp-application-by-using-oauth.md
Configurations described in "Register your application" is not necessary. We cannot use permissions of Microsoft Graph for IMAP, POP and SMTP access. As described in "Get an access token" section, we need to acquire access tokens of Outlook, not Microsoft Graph. And we don't need to add permissions in advance, because we can acquire access tokens dynamically using MSAL or Microsoft Identity Platform v2 endpoint.
And, "Enabled for users to sign-in?" setting of "Office 365 Exchange Online" is "Yes" by default. If this setting is "No", no one can use Exchange Online.

I did a similar document update in the following commit last year.  
https://github.com/MicrosoftDocs/office-developer-exchange-docs/commit/47fc6e17c778a3e809d7daedeb811b707c985eac

Though I couldn't find the exact commit which the unnecessary configurations were added, the following commit is suspicious.  
https://github.com/MicrosoftDocs/office-developer-exchange-docs/commit/99aead68a911ed31f1b1856736892747bf0daed4