How to use the New-ServicePrincipal
I had some problem running the New-ServicePrincipal because I always got the following errors. 

```
new-serviceprincipal : The term 'new-serviceprincipal' is not recognized as the name of a cmdlet, function, script
file, or operable program. Check the spelling of the name, or if a path was included, verify that the path is correct
and try again.
```

The error incorrectly lead me to believe that I had to install some PowerShell module, but it turns out that I missed the appropriate permission in the tenant. I think that it will be useful to give a detailed explanation to people that are not so familiar with the exchange online management module.