Update how-to-create-appointments-in-a-specific-time-zone-by-using-ews-in-exchange.md
When the second example is used to create meeting requests and the recipient is external, the meeting is created fine the first time. If the meeting is update using OWA or EWS (Without outlook touching the meeting first) the TimeZone Description is updated to UTC and causes the meeting to shift unless we stamp the TimeZone Description extended property, the first time correctly. We do have an associated bug open and can provide reference if needed.